Add partial logging + sanitize HTML messages

+++ src/goodreads/goodreads_response.rs
@@ -43,7 +43,7 @@ pub struct ResultsContainer {
 #[derive(Deserialize, Debug, Default)]
 pub struct Work {
     #[serde(deserialize_with = "from_str")]
-    pub id: u32,
+    pub id: u64,
 
     #[serde(deserialize_with = "from_str")]
     pub books_count: u32,
@@ -73,7 +73,7 @@ pub struct Work {
 #[derive(Deserialize, Debug, Default)]
 pub struct Book {
     #[serde(deserialize_with = "from_str")]
-    pub id: u32,
+    pub id: u64,
     pub title: String,
     pub author: Author,
     pub image_url: String,
@@ -83,7 +83,7 @@ pub struct Book {
 #[derive(Deserialize, Debug, Default)]
 pub struct Author {
     #[serde(deserialize_with = "from_str")]
-    pub id: u32,
+    pub id: u64,
     pub name: String,
 }
